refactor: Rename test files module name to tests from previous project output
