+++ tests/read_ascii_test.go
@@ -1,4 +1,4 @@
-package output
+package tests
 
 import (
 	"testing"
